timer symlink ??

+++ BasiliskII/src/SDL/video_sdl2.cpp
@@ -2586,7 +2586,7 @@ static int redraw_func(void *arg)
 
 		// Wait
 		next += VIDEO_REFRESH_DELAY;
-		int32 delay = int32(next - GetTicks_usec());
+		uint64 delay = int32(next - GetTicks_usec());
 		if (delay > 0)
 			Delay_usec(delay);
 		else if (delay < -VIDEO_REFRESH_DELAY)
